Fix scopes
