+++ opencv-traffic-sign/server.py
@@ -1,31 +1,27 @@
+import realtime
 from SocketServer import BaseRequestHandler, UDPServer
 import thread
-import time
+import json
 
-a = 2
+digits = (0)
 
-class TimeHandler(BaseRequestHandler):
+class TrafficSignHandler(BaseRequestHandler):
     def handle(self):
         print('Got connection from', self.client_address)
         # Get message and client socket
         msg, sock = self.request
-        global a
-        resp = str(a)
+        global digits
+        resp = json.dumps(digits)
         sock.sendto(resp, self.client_address)
 
-serv = UDPServer(('', 20000), TimeHandler)
-
-def calculate():	
-    while True:
-        global a
-        a = a + 1
+serv = UDPServer(('', 20001), TrafficSignHandler)
 
 def serve():
     serv.serve_forever()
 
 # Create two threads as follows
 try:
-   thread.start_new_thread(calculate, ())
+   thread.start_new_thread(realtime.calculate, ())
    thread.start_new_thread(serve, ())
 except:
    print "Error: unable to start thread"
